+++ src/ska_mid_cbf_mcs/vcc/vcc_component_manager.py
@@ -82,6 +82,8 @@ class VccComponentManager(CbfObsComponentManager):
 
         self._vcc_controller_proxy = None
 
+        self.last_hps_scan_configuration = ""
+
         # --- Simulators --- #
         self._band_simulators = [
             VccBandSimulator(vcc_band[0]),
@@ -151,6 +153,7 @@ class VccComponentManager(CbfObsComponentManager):
         self._freq_band_name = ""
         self.config_id = ""
         self.scan_id = 0
+        self.last_hps_scan_configuration = ""
 
     def _load_internal_params(
         self: VccComponentManager,
@@ -392,6 +395,7 @@ class VccComponentManager(CbfObsComponentManager):
             return
 
         # Send the ConfigureScan command to the HPS
+        self.last_hps_scan_configuration = json.dumps(configuration)
         fb_index = self._freq_band_index[self._freq_band_name]
 
         if self.simulation_mode: