+++ tests/unit/vcc/Vcc_test.py
@@ -314,12 +314,41 @@ class TestVcc:
         command_dict["ConfigureScan"] = device_under_test.ConfigureScan(
             json_str
         )
+
+        # Test last input scan configuration is correct, must be done before GoToIdle
+        command_name, return_value = (
+            "ConfigureScan",
+            command_dict["ConfigureScan"],
+        )
+        assert return_value[0] == ResultCode.QUEUED
+
+        # Check that the queued command succeeded
+        assert_that(event_tracer).within_timeout(
+            test_utils.EVENT_TIMEOUT
+        ).has_change_event_occurred(
+            device_name=device_under_test,
+            attribute_name="longRunningCommandResult",
+            attribute_value=(
+                f"{return_value[1][0]}",
+                f'[{ResultCode.OK.value}, "{command_name} completed OK"]',
+            ),
+        )
+
+        assert device_under_test.lastScanConfiguration == json_str
+
+        configuration["expected_dish_id"] = device_under_test.dishID
+        assert device_under_test.lastHpsScanConfiguration == json.dumps(
+            configuration
+        )
+
         command_dict["Scan"] = device_under_test.Scan(scan_id)
         command_dict["EndScan"] = device_under_test.EndScan()
         command_dict["GoToIdle"] = device_under_test.GoToIdle()
 
         # Assertions for all issued LRC
         for command_name, return_value in command_dict.items():
+            if command_name == "ConfigureScan":
+                pass
             # Check that the command was successfully queued
             assert return_value[0] == ResultCode.QUEUED
 
@@ -334,6 +363,8 @@ class TestVcc:
                     f'[{ResultCode.OK.value}, "{command_name} completed OK"]',
                 ),
             )
+        assert device_under_test.lastScanConfiguration == ""
+        assert device_under_test.lastHpsScanConfiguration == ""
 
         # Check all obsState transitions
         previous_state = ObsState.IDLE